feat: close cross-language feature parity gaps (C, Rust, Python, TS)
C:
- Implement gzip/zstd framed encode/decode (was stub)
- Add MAX_EXT_LEN, MAX_RANK, MAX_DICT_LEN security limits
- Link zlib required, zstd optional (COWRIE_HAS_ZSTD)

Rust:
- Add typed ImageFormat/AudioEncoding enums (was raw u8)
- Add MAX_EXT_LEN, MAX_DICT_LEN limits, enforce MAX_RANK on tensor decode
- Named JSON serialization for image/audio formats

Python:
- Add DecodeOptions dataclass for configurable limits
- Add decompression bomb protection (256MB limit)
- Add MAX_DICT_LEN check, thread opts through all decode paths

TypeScript:
- Add zstd decompression via fzstd (optional dependency)
- Add MAX_DICT_LEN, MAX_DECOMPRESSED_SIZE limits
- Tighten decodeFramed default from 1GB to 256MB

+++ rust/src/gen2/encode.rs
@@ -253,7 +253,7 @@ fn encode_value(buf: &mut Vec<u8>, value: &Value, dict: &HashMap<&str, usize>, o
         }
         Value::Image(img) => {
             buf.push(tags::IMAGE);
-            buf.push(img.format);
+            buf.push(img.format as u8);
             buf.extend_from_slice(&img.width.to_le_bytes());
             buf.extend_from_slice(&img.height.to_le_bytes());
             write_uvarint(buf, img.data.len() as u64);
@@ -261,7 +261,7 @@ fn encode_value(buf: &mut Vec<u8>, value: &Value, dict: &HashMap<&str, usize>, o
         }
         Value::Audio(aud) => {
             buf.push(tags::AUDIO);
-            buf.push(aud.encoding);
+            buf.push(aud.encoding as u8);
             buf.extend_from_slice(&aud.sample_rate.to_le_bytes());
             buf.push(aud.channels);
             write_uvarint(buf, aud.data.len() as u64);
